\name{univOutl-package}
\alias{univOutl-package}
\alias{univOutl}
\docType{package}
\title{
Detection of Univariate Outliers.
}
\description{
Well known outlier detection techniques in the univariate case. Ratios of two variables are covered too. When available, survey weights can be considered.
}
\details{
The package provides few simple functions implementing well known outlier detection techniques in the univariate case. 
Methods to deal with skewed distribution are included.
The Hidiroglou-Berthelot (1986) method to search for outliers in ratios of historical data is implemented as well. When available, survey weights can be used in outliers detection.
}
\author{
Author and Maintainer: Marcello D'Orazio \email{mdo.statmatch@gmail.com}
}
\references{
Hidiroglou, M.A. and Berthelot, J.-M. (1986) `Statistical editing and Imputation for Periodic Business Surveys'. \emph{Survey Methodology}, Vol 12, pp. 73-83.

McGill, R., Tukey, J. W. and Larsen, W. A. (1978) `Variations of box plots'. \emph{The American Statistician}, 32, pp. 12-16.

Rousseeuw, P.J. and Croux, C. (1993) `Alternatives to the Median Absolute Deviation', \emph{Journal of the American Statistical Association} 88, pp. 1273-1283.

Vanderviere, E. and Huber, M. (2008) `An Adjusted Boxplot for Skewed Distributions', \emph{Computational Statistics & Data Analysis}, 52, pp. 5186-5201

}

\keyword{ package }
